@@ -1,6 +1,8 @@
 package org.example;
 
-import org.example.Data.TokenClass;
+
+import okhttp3.*;
+import org.example.data.BotData;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.api.methods.send.SendDice;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
@@ -12,31 +14,36 @@ import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.Keyboard
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.KeyboardRow;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 public class Bot extends TelegramLongPollingBot {
 
-    private static final String TOKEN = TokenClass.TOKEN;
+    private static final String TOKEN = BotData.TOKEN;
 
     private final ExecutorService executorService;
 
+    private final OkHttpClient okHttpClient = new OkHttpClient();
 
     KeyboardButton spin = KeyboardButton.builder()
             .text("Крутить")
             .build();
 
+    KeyboardButton stat = KeyboardButton.builder()
+            .text("Таблица рекордов")
+            .build();
 
 
     ReplyKeyboardMarkup keyboard = ReplyKeyboardMarkup.builder()
-            .keyboardRow(new KeyboardRow(List.of(spin)))
+            .keyboardRow(new KeyboardRow(List.of(spin, stat)))
             .resizeKeyboard(true)
             .build();
 
     public Bot() {
         super(TOKEN);
-        this.executorService = Executors.newFixedThreadPool(20);
+        this.executorService = Executors.newFixedThreadPool(200);
     }
 
     @Override
@@ -44,16 +51,50 @@ public class Bot extends TelegramLongPollingBot {
         executorService.execute(() -> {
             Message message = update.getMessage();
             Long who = message.getFrom().getId();
+            String username = message.getFrom().getUserName();
             String what = message.getText();
             if (message.isCommand()){
                 if (what.equals("/start")) {
+                    try {
+                        addUser(username);
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
                     onStart(who);
                 }
             } else if (what.equals("Крутить")){
-                System.out.print(message.getFrom().getUserName());
-                spinRoulette(who);
+                System.out.print(username);
+                spinRoulette(who, username);
+            } else if (what.equals("Таблица рекордов")) {
+                try {
+                    sendMessage(who, getScoreTable(username));
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
             }
         });
+    }
+
+
+    public void addUser(String username) throws IOException {
+        RequestBody requestBody = new FormBody.Builder()
+                .add("username", username)
+                .build();
+        Request request = new Request.Builder()
+                .url("http://localhost:8080/api/users/addUser")
+                .post(requestBody)
+                .build();
+        okHttpClient.newCall(request).execute();
+    }
+
+    public String getScoreTable(String username) throws IOException {
+            Request request = new Request.Builder()
+                    .get()
+                    .url("http://localhost:8080/api/users/findAll?username=" + username)
+                    .build();
+            Response response = okHttpClient.newCall(request).execute();
+            assert response.body() != null;
+            return response.body().string();
     }
 
     @Override
@@ -75,7 +116,7 @@ public class Bot extends TelegramLongPollingBot {
         }
     }
 
-    public void spinRoulette(Long who){
+    public void spinRoulette(Long who, String username){
         executorService.execute(() -> {
             ReplyKeyboardRemove remove = ReplyKeyboardRemove.builder().removeKeyboard(true).build();
             SendDice sticker = SendDice.builder()
@@ -86,32 +127,42 @@ public class Bot extends TelegramLongPollingBot {
             try {
                 Message resultMessage = execute(sticker);
                 int value = resultMessage.getDice().getValue();
-                String message;
-                switch (value){
-                    case 1:
-                        message = "Поздравляем! Коэфициент вашей победы: 1";
-                        break;
-                    case 22:
-                        message = "Поздравляем! Коэфициент вашей победы: 2";
-                        break;
-                    case 43:
-                        message = "Поздравляем! Коэфициент вашей победы: 3";
-                        break;
-                    case 64:
-                        message = "Поздравляем! Коэфициент вашей победы: 4";
-                        break;
-                    default:
-                        message = "Повезет в следующий раз! Додеп и крути!";
-                }
+                String message = switch (value) {
+                    case 1 -> "Поздравляем! Коэфициент вашей победы: 1";
+                    case 22 -> "Поздравляем! Коэфициент вашей победы: 2";
+                    case 43 -> "Поздравляем! Коэфициент вашей победы: 3";
+                    case 64 -> "Поздравляем! Коэфициент вашей победы: 4";
+                    default -> "Повезет в следующий раз! Додеп и крути!";
+                };
                 System.out.println(" " + value + " " +message);
                 try {
                     Thread.sleep(2000);
                 } catch (InterruptedException e) {
                     throw new RuntimeException(e);
                 }
+                addScore(username, value);
                 sendMessage(who, message);
 
-            } catch (TelegramApiException e) {
+            } catch (TelegramApiException | IOException e) {
+                throw new RuntimeException(e);
+            }
+        });
+    }
+
+
+    public void addScore(String username, Integer score) throws IOException {
+        executorService.execute(() -> {
+            RequestBody requestBody = new FormBody.Builder()
+                    .add("username", username)
+                    .add("score", String.valueOf(score))
+                    .build();
+            Request request = new Request.Builder()
+                    .url("http://localhost:8080/api/users/addScore")
+                    .put(requestBody)
+                    .build();
+            try {
+                okHttpClient.newCall(request).execute();
+            } catch (IOException e) {
                 throw new RuntimeException(e);
             }
         });
