@@ -1,5 +1,6 @@
 package org.example;
 
+
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
@@ -10,9 +11,8 @@ import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
  */
 public class App 
 {
-    public static void main( String[] args ) throws TelegramApiException {
+    public static void main(String[] args) throws TelegramApiException {
         TelegramBotsApi api = new TelegramBotsApi(DefaultBotSession.class);
-        Bot bot = new Bot();
-        api.registerBot(bot);
+        api.registerBot(new Bot());
     }
 }
